MTDSA-11611- Rebase changes.
